requirements.txt
amqp==2.5.2
asgiref==3.2.7
Babel==2.8.0
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==3.0.6
django-braces==1.14.0
django-celery-results==1.2.1
django-cors-headers==3.3.0
django-filters==0.2.1
django-phone-field==1.8.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.9
kombu==4.6.8
markup==0.2
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
